baseinitData에_데이터_추가

+++ backend/project1/src/main/java/com/team4/project1/domain/item/controller/ApiV1ItemController.java
@@ -19,7 +19,7 @@ public class ApiV1ItemController {
     private final ItemService itemService;
 
     @GetMapping("/{itemId}")
-    public ResponseEntity<ResponseDto<ItemDto>> item(@PathVariable Integer itemId) {
+    public ResponseEntity<ResponseDto<ItemDto>> item(@PathVariable Long itemId) {
         return ResponseEntity.ok(ResponseDto.ok(
                 itemService.getItemById(itemId).orElseThrow(() -> new ItemNotFoundException(itemId)
                 )
